# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Digital HubX API.",
        "title": "Digital HubX API",
        "contact": {
            "name": "API Support (Tanapong R)",
            "url": "mailto:zTanapongR@pttep.com",
            "email": "zTanapongR@pttep.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:3000",
    "basePath": "/hubx-service",
    "paths": {
...
        "/admin/user-question-answer/{class_id}/class": {
            "post": {
                "description": "Allows an admin to submit multiple question answers for a specific class.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/User Question Answer"
                ],
                "summary": "Create multiple user question answers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Class ID",
                        "name": "class_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "List of user question answers",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.CreateUserQuestionAnswerReq"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response\" example:{\"message\":\"User question answers created successfully\",\"status\":200,\"details\":null}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid input\" example:{\"message\":\"Invalid input\",\"status\":400,\"details\":{\"field\":\"error description\"}}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error\" example:{\"message\":\"Internal Server Error\",\"status\":500,\"details\":null}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/user-question-answer/{class_id}/class": {
            "get": {
                "description": "Fetches a paginated list of user question answers for a specific class.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserQuestionAnswer"
                ],
                "summary": "Retrieve user question answers by class ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Class ID",
                        "name": "class_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page (default: 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response\" example:{\"data\":[...],\"total\":100,\"page\":1,\"limit\":10,\"totalPages\":10}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error\" example:{\"message\":\"Internal Server Error\",\"status\":500,\"details\":null}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.CreateUserQuestionAnswerReq": {
            "type": "object",
            "required": [
                "question_id"
            ],
            "properties": {
                "answer_text": {
                    "type": "string"
                },
                "question_id": {
                    "type": "string"
                },
                "selected_choice_id": {
                    "type": "string"
                },
                "sub_question_answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SubQuestionsAnswer"
                    }
                }
            }
        },
...
        "dto.SubQuestionsAnswer": {
            "description": "Details of a sub-question answer.",
            "type": "object",
            "properties": {
                "answer_text": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "parent_choice_id": {
                    "type": "string"
                },
                "question_type": {
                    "type": "string"
                },
                "selected_sub_question_choice_description": {
                    "type": "string"
                },
                "selected_sub_question_choice_id": {
                    "type": "string"
                },
                "sub_question_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "This security definition is used for authenticating",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}